+++ docs/source/conf.py
@@ -24,9 +24,9 @@ copyright = u'2018, Peter Morawski'
 author = u'Peter Morawski'
 
 # The short X.Y version
-version = u''
+version = u'0.1.8'
 # The full version, including alpha/beta/rc tags
-release = u'0.1'
+release = u'0.1.8-PRE-ALPHA'
 
 
 # -- General configuration ---------------------------------------------------
@@ -41,6 +41,7 @@ release = u'0.1'
 extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.viewcode',
+    'sphinx.ext.autodoc',
 ]
 
 # Add any paths that contain templates here, relative to this directory.